test(mempool): cleanup, move last existing flow test
