+++ crates/mempool/tests/flow_test.rs
@@ -54,6 +54,23 @@ fn test_add_tx_fills_nonce_gap(mut mempool: Mempool) {
 }
 
 #[rstest]
+fn test_add_tx_after_get_txs_fails_on_duplicate_nonce(mut mempool: Mempool) {
+    // Setup.
+    let input_tx = add_tx_input!(tx_hash: 0, tx_nonce: 0);
+
+    // Test.
+    add_tx(&mut mempool, &input_tx);
+    get_txs_and_assert_expected(&mut mempool, 1, &[input_tx.tx]);
+
+    let input_tx_duplicate_nonce = add_tx_input!(tx_hash: 1, tx_nonce: 0);
+    add_tx_expect_error(
+        &mut mempool,
+        &input_tx_duplicate_nonce,
+        MempoolError::DuplicateNonce { address: contract_address!("0x0"), nonce: nonce!(0) },
+    );
+}
+
+#[rstest]
 fn test_add_same_nonce_tx_after_previous_not_included_in_block(mut mempool: Mempool) {
     // Setup.
     let tx_nonce_3_account_nonce_3 =
